+++ Source/Validation/ConversionVal.hpp
@@ -1,5 +1,24 @@
 // © 2021 NVIDIA Corporation
 
+constexpr std::array<const char*, (size_t)DescriptorType::MAX_NUM> g_descriptorTypeNames = {
+    "SAMPLER",                      // SAMPLER
+    "MUTABLE",                      // MUTABLE
+    "TEXTURE",                      // TEXTURE
+    "STORAGE_TEXTURE",              // STORAGE_TEXTURE
+    "INPUT_ATTACHMENT",             // INPUT_ATTACHMENT
+    "BUFFER",                       // BUFFER
+    "STORAGE_BUFFER",               // STORAGE_BUFFER
+    "CONSTANT_BUFFER",              // CONSTANT_BUFFER
+    "STRUCTURED_BUFFER",            // STRUCTURED_BUFFER
+    "STORAGE_STRUCTURED_BUFFER",    // STORAGE_STRUCTURED_BUFFER
+    "ACCELERATION_STRUCTURE",       // ACCELERATION_STRUCTURE
+};
+NRI_VALIDATE_ARRAY_BY_PTR(g_descriptorTypeNames);
+
+const char* nri::GetDescriptorTypeName(DescriptorType descriptorType) {
+    return g_descriptorTypeNames[(uint32_t)descriptorType];
+}
+
 void nri::ConvertBotomLevelGeometries(const BottomLevelGeometryDesc* geometries, uint32_t geometryNum, BottomLevelGeometryDesc*& outGeometries, BottomLevelMicromapDesc*& outMicromaps) {
     for (uint32_t i = 0; i < geometryNum; i++) {
         const BottomLevelGeometryDesc& src = geometries[i];
